+++ src/components/kanban/__tests__/KanbanBoard.test.tsx
@@ -33,6 +33,7 @@ function makeFilters(search = ''): UIContextValue['state']['filters'] {
     stuckDays: null,
     errorState: 'none',
     sortBy: 'updatedAt_desc',
+    showMocks: true,
   }
 }