+++ src/components/layout/__tests__/AppShell.test.tsx
@@ -93,6 +93,7 @@ function makeFilters(search = ''): UIContextValue['state']['filters'] {
     stuckDays: null,
     errorState: 'none',
     sortBy: 'updatedAt_desc',
+    showMocks: true,
   }
 }