+++ src/components/ticket/__tests__/TicketForm.test.tsx
@@ -31,6 +31,7 @@ function makeFilters(): UIContextValue['state']['filters'] {
     stuckDays: null,
     errorState: 'none',
     sortBy: 'updatedAt_desc',
+    showMocks: true,
   }
 }